backend/pyproject.toml
[project]
name = "physical-ai-humanoid-robotics-chatbot"
version = "1.0.0"
description = "Chatbot API for the Physical AI & Humanoid Robotics documentation website"
readme = "README.md"
authors = [
    { name = "Muhammad-Azan-1", email = "azan96593@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "python-dotenv==1.0.1",
    "fastapi[standard]==0.115.0",
    "uvicorn[standard]==0.32.0",
    "redis==5.2.0",
    "pydantic>=2.10.3",
    "openai>=1.55.0",
    "pinecone>=3.1.0",
    "python-multipart==0.0.20",
    "requests>=2.32.3",
    "google-generativeai>=0.8.3",
    "openai-agents>=0.6.3",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.uv.sources]
openai-agents = { git = "https://github.com/openai/openai-agents-python.git" }

